Extract potentially-generic features to `Http4sBaseSuite`
